EN-3199 : change type of parameter round from uint32 to uint64 (module was updated and this change was required)

+++ process/block/export_test.go
@@ -266,7 +266,7 @@ func (sp *shardProcessor) RestoreMetaBlockIntoPool(
 }
 
 func (sp *shardProcessor) GetAllMiniBlockDstMeFromMeta(
-	round uint32,
+	round uint64,
 	metaHashes [][]byte,
 ) (map[string][]byte, error) {
 	return sp.getAllMiniBlockDstMeFromMeta(round, metaHashes)
